test: fix port id type
Fixes: f8244c6399d9 ("ethdev: increase port id range")

+++ test/test-pipeline/init.c
@@ -194,9 +194,9 @@ app_ports_check_link(void)
 
 	for (i = 0; i < app.n_ports; i++) {
 		struct rte_eth_link link;
-		uint8_t port;
+		uint16_t port;
 
-		port = (uint8_t) app.ports[i];
+		port = app.ports[i];
 		memset(&link, 0, sizeof(link));
 		rte_eth_link_get_nowait(port, &link);
 		RTE_LOG(INFO, USER1, "Port %u (%u Gbps) %s\n",
@@ -219,10 +219,10 @@ app_init_ports(void)
 
 	/* Init NIC ports, then start the ports */
 	for (i = 0; i < app.n_ports; i++) {
-		uint8_t port;
+		uint16_t port;
 		int ret;
 
-		port = (uint8_t) app.ports[i];
+		port = app.ports[i];
 		RTE_LOG(INFO, USER1, "Initializing NIC port %u ...\n", port);
 
 		/* Init port */
